git-svn-id: https://localhost/svn/Project64/trunk@32 111125ac-702d-7242-af9c-5ba8ae61c1ef

+++ Source/Project64/N64 System/Recompiler/Exit Info.h
@@ -18,7 +18,7 @@ public:
 	CRegInfo    ExitRegSet;
 	EXIT_REASON reason;
 	STEP_TYPE   NextInstruction;
-	BYTE *      JumpLoc; //32bit jump
+	DWORD *     JumpLoc; //32bit jump
 };
 
 typedef std::list<CExitInfo> EXIT_LIST;
